# apps/runner/pkg/api/docs/swagger.yaml
definitions:
...
  CreateBackupDTO:
    properties:
      registry:
        $ref: '#/definitions/RegistryDTO'
      snapshot:
        type: string
    required:
    - registry
    - snapshot
    type: object
  CreateSandboxDTO:
    properties:
      cpuQuota:
        minimum: 1
        type: integer
      entrypoint:
        items:
          type: string
        type: array
      env:
        additionalProperties:
          type: string
        type: object
      fromVolumeId:
        type: string
      gpuQuota:
        minimum: 0
        type: integer
      id:
        type: string
      memoryQuota:
        minimum: 1
        type: integer
      metadata:
        additionalProperties:
          type: string
        type: object
      networkAllowList:
        type: string
      networkBlockAll:
        type: boolean
      osUser:
        type: string
      registry:
        $ref: '#/definitions/RegistryDTO'
      snapshot:
        type: string
      storageQuota:
        minimum: 1
        type: integer
      userId:
        type: string
      volumes:
        items:
          $ref: '#/definitions/dto.VolumeDTO'
        type: array
    required:
    - id
    - osUser
    - snapshot
    - userId
    type: object
  ErrorResponse:
    description: Error response
    properties:
      code:
        example: BAD_REQUEST
        type: string
      message:
        example: Bad request
        type: string
      method:
        example: GET
        type: string
      path:
        example: /api/resource
        type: string
      statusCode:
        example: 400
        type: integer
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
    required:
    - code
    - message
    - method
    - path
    - statusCode
    - timestamp
    type: object
...
  RegistryDTO:
    properties:
      password:
        type: string
      project:
        type: string
      url:
        type: string
      username:
        type: string
    required:
    - password
    - url
    - username
    type: object
  ResizeSandboxDTO:
    properties:
      cpu:
        minimum: 1
        type: integer
      gpu:
        minimum: 0
        type: integer
      memory:
        minimum: 1
        type: integer
    type: object
...
  SandboxInfoResponse:
    properties:
      backupError:
        type: string
      backupState:
        $ref: '#/definitions/enums.BackupState'
      state:
        $ref: '#/definitions/enums.SandboxState'
    type: object
...
  UpdateNetworkSettingsDTO:
    properties:
      networkAllowList:
        type: string
      networkBlockAll:
        type: boolean
      networkLimitEgress:
        type: boolean
    type: object
  dto.VolumeDTO:
    properties:
      mountPath:
        type: string
      volumeId:
        type: string
    type: object
  enums.BackupState:
    enum:
    - NONE
    - PENDING
    - IN_PROGRESS
    - COMPLETED
    - FAILED
    type: string
    x-enum-varnames:
    - BackupStateNone
    - BackupStatePending
    - BackupStateInProgress
    - BackupStateCompleted
    - BackupStateFailed
  enums.SandboxState:
    enum:
    - creating
    - restoring
    - destroyed
    - destroying
    - started
    - stopped
    - starting
    - stopping
    - resizing
    - error
    - unknown
    - pulling_snapshot
    type: string
    x-enum-varnames:
    - SandboxStateCreating
    - SandboxStateRestoring
    - SandboxStateDestroyed
    - SandboxStateDestroying
    - SandboxStateStarted
    - SandboxStateStopped
    - SandboxStateStarting
    - SandboxStateStopping
    - SandboxStateResizing
    - SandboxStateError
    - SandboxStateUnknown
    - SandboxStatePullingSnapshot
info:
  contact: {}
  description: Daytona Runner API
  title: Daytona Runner API
  version: v0.0.0-dev
paths:
...
  /sandboxes:
    post:
      description: Create a sandbox
      operationId: Create
      parameters:
      - description: Create sandbox
        in: body
        name: sandbox
        required: true
        schema:
          $ref: '#/definitions/CreateSandboxDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Create a sandbox
      tags:
      - sandbox
  /sandboxes/{sandboxId}:
    get:
      description: Get sandbox info
      operationId: Info
      parameters:
      - description: Sandbox ID
        in: path
        name: sandboxId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sandbox info
          schema:
            $ref: '#/definitions/SandboxInfoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get sandbox info
      tags:
      - sandbox
  /sandboxes/{sandboxId}/backup:
    post:
      description: Create sandbox backup
      operationId: CreateBackup
      parameters:
      - description: Sandbox ID
        in: path
        name: sandboxId
        required: true
        type: string
      - description: Create backup
        in: body
        name: sandbox
        required: true
        schema:
          $ref: '#/definitions/CreateBackupDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Backup started
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Create sandbox backup
      tags:
      - sandbox
  /sandboxes/{sandboxId}/destroy:
    post:
      description: Destroy sandbox
      operationId: Destroy
      parameters:
      - description: Sandbox ID
        in: path
        name: sandboxId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sandbox destroyed
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Destroy sandbox
      tags:
      - sandbox
  /sandboxes/{sandboxId}/network-settings:
    get:
      description: Get sandbox network settings
      operationId: GetNetworkSettings
      parameters:
      - description: Sandbox ID
        in: path
        name: sandboxId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Network settings
          schema:
            $ref: '#/definitions/UpdateNetworkSettingsDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get sandbox network settings
      tags:
      - sandbox
    post:
      description: Update sandbox network settings
      operationId: UpdateNetworkSettings
      parameters:
      - description: Sandbox ID
        in: path
        name: sandboxId
        required: true
        type: string
      - description: Update network settings
        in: body
        name: sandbox
        required: true
        schema:
          $ref: '#/definitions/UpdateNetworkSettingsDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Network settings updated
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Update sandbox network settings
      tags:
      - sandbox
  /sandboxes/{sandboxId}/resize:
    post:
      description: Resize sandbox
      operationId: Resize
      parameters:
      - description: Sandbox ID
        in: path
        name: sandboxId
        required: true
        type: string
      - description: Resize sandbox
        in: body
        name: sandbox
        required: true
        schema:
          $ref: '#/definitions/ResizeSandboxDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Sandbox resized
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Resize sandbox
      tags:
      - sandbox
  /sandboxes/{sandboxId}/start:
    post:
      description: Start sandbox
      operationId: Start
      parameters:
      - description: Sandbox ID
        in: path
        name: sandboxId
        required: true
        type: string
      - description: Metadata
        in: body
        name: metadata
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Sandbox started
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Start sandbox
      tags:
      - sandbox
  /sandboxes/{sandboxId}/stop:
    post:
      description: Stop sandbox
      operationId: Stop
      parameters:
      - description: Sandbox ID
        in: path
        name: sandboxId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sandbox stopped
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Stop sandbox
      tags:
      - sandbox
...
security:
- Bearer: []
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and an API token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"